+++ docs/conf.py
@@ -57,7 +57,7 @@ copyright = '2016, ' + author
 # built documents.
 #
 # The short X.Y version.
-version = '0.1.5'
+version = '0.2.0'
 # The full version, including alpha/beta/rc tags.
 release = version + '-dev'